requirements.txt
beautifulsoup4==4.8.0
decorator==4.4.0
django==5.1.6
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
ipython==8.10.0
ipython-genutils==0.2.0
itypes==1.2.0
lxml==5.3.1
numpy==1.26.4
ofxparse==0.20
pandas==2.2.3
PasteDeploy==2.0.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==3.0.50
psycopg[binary]==3.2.5
ptyprocess==0.6.0
pygments==2.7.4
PyJWT==2.10.1
python-dateutil==2.9.0
pytz==2025.1
PyYAML==6.0.2
requests==2.22.0
scikit-learn==1.6.1
scipy==1.15.2
simplegeneric==0.8.1
six==1.17.0
traitlets==5.14.3
uritemplate==3.0.0
wcwidth==0.1.7
